third_party/blink/web_tests/external/wpt/secure-payment-confirmation/enrollment-in-iframe.sub.https-expected.txt

This is a testharness.js-based test.
[FAIL] SPC enrollment in cross-origin iframe
  assert_equals: expected null but got object "SecurityError: A user activation is required to create a credential in a cross-origin iframe."
[PASS] SPC enrollment in cross-origin iframe fails without user activation
[PASS] SPC enrollment in cross-origin iframe without payment permission
Harness: the test ran to completion.
